refactor: JobItemAdd 리팩토링
- JobAdd -> JobItemAdd 이름 변경
- resume store 사용
- 알림 스낵바 페이지에서 컴포넌트 내로 이동
